requirements.txt
# LiveKit Agent dependencies
livekit-agents==1.2.5
livekit==1.0.12
livekit-api==1.0.5
livekit-protocol==1.0.5

# LiveKit plugins
livekit-plugins-openai==1.2.5
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-turn-detector==1.2.5
livekit-plugins-silero==1.2.5

# Socket.IO TTS dependencies
python-socketio[asyncio]>=5.11.0
aiohttp>=3.9.0

# Other dependencies
python-dotenv>=1.0.0
loguru>=0.7.0

# OpenAI for STT/LLM
openai>=1.84.0

fastapi 
uvicorn 
pyjwt 

